# server/docs/swagger.yaml
basePath: /v1
definitions:
  handler.CreateAddressReq:
    properties:
      city:
        type: string
      detail:
        type: string
      district:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      phone:
        type: string
      province:
        type: string
    required:
    - city
    - detail
    - district
    - name
    - phone
    - province
    type: object
...
  handler.CreateTradeReq:
    properties:
      barter_coin_amount:
        type: number
      message:
        type: string
      offered_item_id:
        type: integer
      target_item_id:
        type: integer
    required:
    - target_item_id
    type: object
...
  handler.RejectReq:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
  model.Category:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
    type: object
...
  model.Item:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      condition:
        type: string
      created_at:
        type: string
      description:
        type: string
      estimated_value:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      location:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 关联
      user_id:
        type: integer
      video_url:
        type: string
      view_count:
        type: integer
      want_items:
        items:
          type: string
        type: array
    type: object
...
  model.TradeRequest:
    properties:
      barter_coin_amount:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      expired_at:
        type: string
      id:
        type: integer
      initiator:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 关联
      initiator_id:
        type: integer
      message:
        type: string
      offered_item:
        $ref: '#/definitions/model.Item'
      offered_item_id:
        type: integer
      reject_reason:
        type: string
      status:
        type: string
      target_item:
        $ref: '#/definitions/model.Item'
      target_item_id:
        type: integer
      target_user:
        $ref: '#/definitions/model.User'
      target_user_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      credit_score:
        type: integer
      enterprise_name:
        type: string
      enterprise_verified:
        type: boolean
      id:
        type: integer
      last_login_at:
        type: string
      location:
        type: string
      nickname:
        type: string
      phone:
        type: string
      real_name_verified:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
    type: object
  model.UserAddress:
    properties:
      city:
        type: string
      created_at:
        type: string
      detail:
        type: string
      district:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      phone:
        type: string
      province:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.Wallet:
    properties:
      balance:
        type: string
      created_at:
        type: string
      frozen_balance:
        type: string
      id:
        type: integer
      total_expense:
        type: string
      total_income:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: NeoBarter Team
    url: https://github.com/sxc19960907/neobarter
  description: 现代易货交易平台后端 API 文档。支持 C2C/B2B 双模式、巴特币结算、智能匹配。
  license:
    name: MIT
  termsOfService: https://github.com/sxc19960907/neobarter
  title: 新易巴特 NeoBarter API
  version: "1.0"
paths:
...
  /trades:
    get:
      parameters:
      - description: 状态筛选：pending/accepted/rejected/completed/cancelled
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取我的交易列表
      tags:
      - 交易
    post:
      consumes:
      - application/json
      description: 对目标物品发起交换，可附带巴特币补差价
      parameters:
      - description: 交换请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTradeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TradeRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 发起交换请求
      tags:
      - 交易
  /trades/{id}:
    get:
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TradeRequest'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取交易详情
      tags:
      - 交易
  /trades/{id}/accept:
    put:
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 接受交换请求
      tags:
      - 交易
  /trades/{id}/cancel:
    put:
      description: 仅发起方可取消待处理的交易
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 取消交易请求
      tags:
      - 交易
  /trades/{id}/complete:
    put:
      description: 双方确认后结算巴特币，物品标记为已交易
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 确认完成交易
      tags:
      - 交易
  /trades/{id}/reject:
    put:
      consumes:
      - application/json
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      - description: 拒绝原因
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RejectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 拒绝交换请求
      tags:
      - 交易
  /upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传物品图片或头像，返回可访问 URL（jpg/png/webp/gif，≤5MB）
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传图片
      tags:
      - 上传
...
  /users/me/addresses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserAddress'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取收货地址列表
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: 地址信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAddressReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserAddress'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 新增收货地址
      tags:
      - 用户
  /users/me/addresses/{id}:
    delete:
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除收货地址
      tags:
      - 用户
    put:
      consumes:
      - application/json
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      - description: 地址信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAddressReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserAddress'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新收货地址
      tags:
      - 用户
  /wallet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Wallet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取巴特币钱包
      tags:
      - 钱包
  /wallet/transactions:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取巴特币流水
      tags:
      - 钱包
securityDefinitions:
  BearerAuth:
    description: 输入格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"